continued refactoring work
