+++ assets/javascript/app.js
@@ -2,8 +2,6 @@ $(document).ready(function(){
   onPageLoadInitialize();
   console.log('Initiated!');
 
-  let elementGeneration = $('#elementGeneration');
-
   let myCodeMirror = CodeMirror(document.getElementById('elementTree'), {
     value: '<!DOCTYPE html>\n' +
     '<html lang="en">\n' +
@@ -21,54 +19,6 @@ $(document).ready(function(){
     tabSize: 2,
     lineNumbers: true
   });
-
-  // ============= Speech Recognition Functions ============= \\
-  function toggleStartStop() {
-    let chatTextArea = $('#chatTextArea');
-    let micIcon = $('#microphoneIcon');
-    let interimTextDisplay = $('#interimTextDisplay');
-    let transcriptDisplay = $('#interimTextDisplayLabel');
-
-    if (recognizing === true) {
-      recognition.stop();
-      micIcon.attr("src", "assets/images/micOff.png");
-      recognizing = false;
-    }
-    else {
-      recognition = new SpeechRecognition();
-      recognition.continuous = true;
-      recognition.interimResults = true;
-      recognition.start();
-      micIcon.attr("src", "assets/images/micOn.png");
-      recognizing = true;
-      recognition.onend = function() {
-        recognizing = false;
-        micIcon.attr("src", "assets/images/micOff.png");
-      };
-      recognition.onresult = function(e) {
-        final = '';
-        interim = '';
-        for (let i = 0; i < e.results.length; ++i) {
-          if (e.results[i].isFinal) {
-            final += e.results[i][0].transcript;
-            console.log('final transcription:', e.results[i][0].transcript);
-            chatTextArea.focus();
-            if (recognizing === true) {
-              toggleStartStop();
-              transcriptDisplay.toggle();
-            }
-          } else {
-            interim += e.results[i][0].transcript;
-          }
-        }
-        chatTextArea.text(final);
-        interimTextDisplay.text(interim);
-      };
-      chatTextArea.text('');
-      interimTextDisplay.text('');
-    }
-  }
-  // ============= END OF: Speech Recognition Functions ============= \\
 
   // click handler to start or stop voice-to-text dictation
   $('#dictationButton').on('click', function(){
@@ -98,7 +48,6 @@ $(document).ready(function(){
       selectText(clickedElement);
     }
   });
-
 
   // FUNCTION FOR EXPANDING CODE-CHECK VALIDATION RESULTS BAR
   $('#validatorInstructions').click(function() {
@@ -205,22 +154,6 @@ function checkSubmittedTextForElementPostSpellCheck() {
     $('#elementGeneration').append(newElementSpan, '\n');
     selectText(newElementSpanId);
     elementCreatedCounter++
-  }
-}
-
-// selects text of item based on supplied element ID
-function selectText(element) {
-  let doc = document, text = doc.getElementById(element), range, selection;
-  if (doc.body.createTextRange) {
-    range = document.body.createTextRange();
-    range.moveToElementText(text);
-    range.select();
-  } else if (window.getSelection) {
-    selection = window.getSelection();
-    range = document.createRange();
-    range.selectNodeContents(text);
-    selection.removeAllRanges();
-    selection.addRange(range);
   }
 }
 
@@ -350,121 +283,3 @@ function checkSubmittedTextForClass() {
   return classResponse
 }
 // ============= END OF: Presence of Class attribute Validation Code ============= \\
-
-// ============= Emailing Element Tree Function ============ \\
-
-function sendEmail(){
-  let service_id = 'yahoo';
-  let template_id = 'template_ZHevUYdN';
-  let emailBody = getElementTreeText();
-  let email = $('#email').val();
-  let emailSubject = $('#emailSubject').val();
-  let template_params = {
-    subject: emailSubject,
-    name: 'Code-Dictator',
-    reply_email: email,
-    message: emailBody
-  };
-  let respond = emailjs.send(service_id,template_id,template_params);
-  console.log(respond);
-  console.log(template_params.subject);
-  console.log(emailBody);
-}
-
-// ============= W3C Validator Code ============= \\
-function codeValidator () {
-  let data = getElementTreeText();
-  let formData = new FormData();
-  console.log(data);
-  formData.append('out', 'json');
-  formData.append('content', elementTreeData);
-
-  let queryURL = 'https://validator.w3.org/nu/';
-
-  // Performing our AJAX GET request
-  $.ajax({
-    url: queryURL,
-    data: formData,
-    dataType: 'json',
-    type: "POST",
-    processData: false,
-    contentType: false,
-    success: function (response) {
-      console.log(response);
-      if (response.messages.length > 0) {
-        console.log('OH OH! Errors found!!!');
-        $('#validationText').html('<p>I found some issues with your code :(</p>');
-        for (let i = 0; i < response.messages.length; i++) {
-          let errorList = $('#list');
-          let errorListItem = $('<li>');
-          let issueTypeSpan = $('<span>');
-          let lineNumberSpan = $('<span>').css('font-weight', 'bold');
-          let extractSpan = $('<span>').css('font-weight', 'bold');
-          let extractSpanText = $('<span>');
-          let subType = response.messages[i].subType; // these are for warnings
-          let lineNumber = response.messages[i].lastLine; // line number issue is on
-          let extract = response.messages[i].extract; // piece of code containing issue
-          let type = response.messages[i].type; // these are for errors
-          lineNumberSpan.text('Line: ');
-          extractSpan.text('Code: ');
-          extractSpanText.text(extract);
-          issueTypeSpan.css({'font-weight': 'bold', 'padding': '5px', 'border-radius': '20px'});
-          if (subType === 'warning') {
-            issueTypeSpan.css('background-color', '#fecc7d').text(subType);
-          }
-          else if (type === 'error') {
-            issueTypeSpan.css('background-color', '#FF908D').text(type);
-          }
-          errorListItem.text(response.messages[i].message)
-            .prepend(issueTypeSpan)
-            .append('<br>', lineNumberSpan, lineNumber, '<br>', extractSpan, extractSpanText);
-          errorList.append(errorListItem, '<br>');
-        }
-      }
-      else {
-        console.log('No Errors Found');
-        $('#validationText').text('No issues found, good job! :)');
-      }
-    },
-    error: function () {
-      console.warn(arguments);
-    }
-  });
-}
-
-// ============= END OF: W3C Validator Code ============= \\
-
-function spellCheck () {
-  let queryURL = "https://api.cognitive.microsoft.com/bing/v7.0/spellcheck?text=" + submittedChat;
-
-  // Performing our AJAX GET request
-  $.ajax({
-    url: queryURL,
-    method: "GET",
-    headers: {
-      "Ocp-Apim-Subscription-Key": "c1e0f9b9a99347a6b0c83b5df9995731",
-      "Accept": "application/json",
-    },
-  })
-  // After the data comes back from the API ...
-  .then(function (response) {
-    console.log(response);
-    // IF errors were found...
-    if (response.flaggedTokens.length > 0) {
-      // console.log('Before fixing mistakes (if any):', sentenceArray);
-      for (let i = 0; i < response.flaggedTokens.length; i++) {
-        sentenceArray[sentenceArray.indexOf(response.flaggedTokens[i].token)] = response.flaggedTokens[i].suggestions[0].suggestion;
-      }
-      // console.log('After fixing mistakes:', sentenceArray);
-      let message = computerMessages.spellingCorrected + sentenceArray.join(' ') + '"';
-      createChatLineItem('The Dictator', message);
-      knownCommandsCheck();
-      recognition = '';
-    }
-    else {
-      console.log('Spelling validation passed!');
-      knownCommandsCheck();
-      recognition = '';
-    }
-  });
-}
