+++ assets/javascript/globalAssets.js
@@ -79,3 +79,19 @@ function getElementTreeText() {
   });
   return elementTreeData;
 }
+
+// selects text of item based on supplied element ID
+function selectText(element) {
+  let doc = document, text = doc.getElementById(element), range, selection;
+  if (doc.body.createTextRange) {
+    range = document.body.createTextRange();
+    range.moveToElementText(text);
+    range.select();
+  } else if (window.getSelection) {
+    selection = window.getSelection();
+    range = document.createRange();
+    range.selectNodeContents(text);
+    selection.removeAllRanges();
+    selection.addRange(range);
+  }
+}
