--- assets/javascript/ulil.js
@@ -1,9 +1,0 @@
-
-export const returnVal = ()=>{
-  return 'val'
-};
-
-export const concatVal = (arg)=>{
-  console.log(arg);
-  return `this ${arg}`
-};
