+++ index.html
@@ -17,6 +17,11 @@
   <script src="assets/javascript/elementsLibrary.js"></script>
   <script src="assets/javascript/aiResponseLibrary.js"></script>
   <script src="assets/javascript/commandsLibrary.js"></script>
+  <script src="assets/javascript/speechRecognition.js"></script>
+  <script src="assets/javascript/codeValidate.js"></script>
+  <script src="assets/javascript/emailElementTree.js"></script>
+  <script src="assets/javascript/spellCheck.js"></script>
+
   <script src="assets/javascript/app.js"></script>
 
   <!--LINKING MOMENT.JS-->
